users/requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
gunicorn
requests
django-rest-swagger
